METAGEN-9 Extended infrastructure to allow the specification of persistence.xml

+++ tooling/metamodel-generator/src/test/java/org/hibernate/jpamodelgen/test/util/TestUtil.java
@@ -18,11 +18,14 @@
 package org.hibernate.jpamodelgen.test.util;
 
 import java.io.File;
+import java.io.FileFilter;
 import java.lang.reflect.Field;
 import java.lang.reflect.GenericArrayType;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.testng.Assert;
 
 import static org.testng.Assert.assertEquals;
@@ -34,8 +37,9 @@ import static org.testng.FileAssert.fail;
  * @author Hardy Ferentschik
  */
 public class TestUtil {
-
+	private static final Logger log = LoggerFactory.getLogger( TestUtil.class );
 	private static final String PATH_SEPARATOR = System.getProperty( "file.separator" );
+	private static final String PACKAGE_SEPARATOR = ".";
 	private static final String META_MODEL_CLASS_POSTFIX = "_";
 	private static final String outBaseDir;
 
@@ -50,11 +54,23 @@ public class TestUtil {
 	private TestUtil() {
 	}
 
-	public static void clearOutputFolder() {
-		File outDir = new File( outBaseDir );
-		File[] files = outDir.listFiles();
-		for ( File file : files ) {
-			file.delete();
+	public static void deleteGeneratedSourceFiles(File path) {
+		if ( path.exists() ) {
+			File[] files = path.listFiles( new MetaModelFilenameFilter() );
+			for ( File file : files ) {
+				if ( file.isDirectory() ) {
+					deleteGeneratedSourceFiles( file );
+				}
+				else {
+					boolean success = file.delete();
+					if ( success ) {
+						log.debug( file.getAbsolutePath() + " deleted successfully" );
+					}
+					else {
+						log.debug( "Failed to delete generated source file" + file.getAbsolutePath() );
+					}
+				}
+			}
 		}
 	}
 
@@ -78,7 +94,7 @@ public class TestUtil {
 		assertNotNull( clazz, "Class parameter cannot be null" );
 		String metaModelClassName = clazz.getName() + META_MODEL_CLASS_POSTFIX;
 		// generate the file name
-		String fileName = metaModelClassName.replace( ".", PATH_SEPARATOR );
+		String fileName = metaModelClassName.replace( PACKAGE_SEPARATOR, PATH_SEPARATOR );
 		fileName = fileName.concat( ".java" );
 		File sourceFile = new File( outBaseDir + PATH_SEPARATOR + fileName );
 		assertFalse( sourceFile.exists(), "There should be no source file: " + fileName );
@@ -98,14 +114,18 @@ public class TestUtil {
 
 	public static void assertAttributeTypeInMetaModelFor(Class<?> clazz, String fieldName, Class<?> expectedType, String errorString) {
 		Field field = getFieldFromMetamodelFor( clazz, fieldName );
-		assertNotNull( field );
+		assertNotNull( field, "Cannot find field '" + fieldName + "' in " + clazz.getName() );
 		ParameterizedType type = ( ParameterizedType ) field.getGenericType();
 		Type actualType = type.getActualTypeArguments()[1];
 		if ( expectedType.isArray() ) {
 			expectedType = expectedType.getComponentType();
 			actualType = ( ( GenericArrayType ) actualType ).getGenericComponentType();
 		}
-		assertEquals( actualType, expectedType, errorString );
+		assertEquals(
+				actualType,
+				expectedType,
+				"Types do not match: " + errorString
+		);
 	}
 
 	public static void assertMapAttributesInMetaModelFor(Class<?> clazz, String fieldName, Class<?> expectedMapKey, Class<?> expectedMapValue, String errorString) {
@@ -156,6 +176,23 @@ public class TestUtil {
 			field = null;
 		}
 		return field;
+	}
+
+	public static String fcnToPath(String fcn) {
+		return fcn.replace( PACKAGE_SEPARATOR, PATH_SEPARATOR );
+	}
+
+	private static class MetaModelFilenameFilter implements FileFilter {
+		@Override
+		public boolean accept(File pathName) {
+			if ( pathName.isDirectory() ) {
+				return true;
+			}
+			else {
+				return pathName.getAbsolutePath().endsWith( "_.java" )
+						|| pathName.getAbsolutePath().endsWith( "_.class" );
+			}
+		}
 	}
 }
